+++ demo/src/main/java/login/permission/project/classes/DemoApplication.java
@@ -1,4 +1,4 @@
-package classes;
+package login.permission.project.classes;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
