+++ demo/src/main/java/login/permission/project/classes/model/Employee.java
@@ -1,4 +1,4 @@
-package classes;
+package login.permission.project.classes.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
