+++ demo/src/main/java/login/permission/project/classes/repository/DepartmentRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package login.permission.project.classes.repository;
 
-import classes.Department;
+import login.permission.project.classes.model.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DepartmentRepository extends JpaRepository<Department,Integer> {
