+++ demo/src/main/java/login/permission/project/classes/repository/LoginRecordRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package login.permission.project.classes.repository;
 
-import classes.LoginRecord;
+import login.permission.project.classes.model.LoginRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LoginRecordRepository extends JpaRepository<LoginRecord,Integer> {
