+++ demo/src/main/java/login/permission/project/classes/repository/PermissionMapRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package login.permission.project.classes.repository;
 
-import classes.PermissionMap;
+import login.permission.project.classes.model.PermissionMap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PermissionMapRepository extends JpaRepository<PermissionMap,Integer> {
