+++ demo/src/main/java/login/permission/project/classes/repository/PermissionRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package login.permission.project.classes.repository;
 
-import classes.Permission;
+import login.permission.project.classes.model.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PermissionRepository extends JpaRepository<Permission,Integer> {
