+++ demo/src/main/java/login/permission/project/classes/repository/UnitRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package login.permission.project.classes.repository;
 
-import classes.Unit;
+import login.permission.project.classes.model.Unit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UnitRepository extends JpaRepository<Unit,Integer> {
