+++ demo/src/main/java/login/permission/project/classes/service/DatabaseService.java
@@ -1,0 +1,25 @@
+package login.permission.project.classes.service;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+@Service
+public class DatabaseService {
+
+  @Value("${spring.datasource.url}")
+  private String datasourceUrl;
+
+  @Value("${spring.datasource.username}")
+  private String username;
+
+  @Value("${spring.datasource.password}")
+  private String password;
+
+  public Connection connect() throws SQLException {
+    return DriverManager.getConnection(datasourceUrl, username, password);
+  }
+}
