+++ demo/src/main/resources/application.properties
@@ -1,3 +1,4 @@
+server.port=8085
 spring.application.name=demo
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://fengjiapark-a254-1.tc.kingnet.net.tw:11699/LoginPermissionDataBase
@@ -9,3 +10,4 @@ spring.jpa.properties.hibernate.dialect.storage_engine=innodb
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
+spring.contextPath=/CheckListMavenWebThree
